refactor: put FileId to store-api
