+++ src/mito2/src/compaction/buckets.rs
@@ -80,9 +80,10 @@ pub(crate) const TIME_BUCKETS: TimeBuckets = TimeBuckets([
 
 #[cfg(test)]
 mod tests {
+    use store_api::storage::FileId;
+
     use super::*;
     use crate::compaction::test_util::new_file_handle;
-    use crate::sst::file::FileId;
 
     #[test]
     fn test_time_bucket() {
