+++ src/mito2/src/manifest/tests/checkpoint.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use common_datasource::compression::CompressionType;
-use store_api::storage::RegionId;
+use store_api::storage::{FileId, RegionId};
 use strum::IntoEnumIterator;
 
 use crate::error::Error::ChecksumMismatch;
@@ -28,7 +28,7 @@ use crate::manifest::manager::RegionManifestManager;
 use crate::manifest::storage::CheckpointMetadata;
 use crate::manifest::tests::utils::basic_region_metadata;
 use crate::region::{RegionLeaderState, RegionRoleState};
-use crate::sst::file::{FileId, FileMeta};
+use crate::sst::file::FileMeta;
 use crate::test_util::TestEnv;
 
 async fn build_manager(
