+++ src/store-api/src/storage.rs
@@ -16,6 +16,7 @@
 
 pub mod consts;
 mod descriptors;
+mod file;
 mod requests;
 mod types;
 
@@ -25,5 +26,6 @@ pub use datatypes::schema::{
 };
 
 pub use self::descriptors::*;
+pub use self::file::{FileId, ParseIdError};
 pub use self::requests::{ScanRequest, TimeSeriesDistribution, TimeSeriesRowSelector};
 pub use self::types::SequenceNumber;
